refactor: cleanup & simplify
